Move provider scripts into cmd directory
All entry scripts will be under cmd.
