+++ keylime/cmd/ima_emulator_adapter.py
