+++ keylime/cmd/provider_registrar.py
