+++ keylime/cmd/provider_vtpm_add.py
